BUNCH-83: add RecordTravelUpdate helper, 6 travel Apply methods, dispatch registration, and TDD tests
Apply tests pass (6/6). Replay-equality tests are red (expected) until
command methods are migrated to produce events in Tasks 6-9. Heat deltas
corrected to int to match domain (PursuitState.Heat is int).

+++ src/WildBunch.Domain/Events/JourneyEncounterResolved.cs
@@ -21,7 +21,7 @@ public sealed record JourneyEncounterResolved : IDomainEvent
     public required int AmmoSpent { get; init; }
     public required ItemKind? StolenItemKind { get; init; }
     public required int StolenItemQuantity { get; init; }
-    public required decimal PursuitHeatDelta { get; init; }
+    public required int PursuitHeatDelta { get; init; }
     public required int HorseExhaustionDelta { get; init; }
     public required bool ContinuedOnFoot { get; init; }
     public required TravelJourneySnapshot JourneySnapshot { get; init; }
